Migrating Shop & CPUCooler

+++ Computer Era X/DataTypes/Interfaces/IItem.cs
@@ -7,7 +7,7 @@ namespace Computer_Era_X.DataTypes.Interfaces
         int ID { get; set; }
         string Name { get; set; }
         string Type { get; set; }
-        int Price { get; set; }
+        double Price { get; set; }
         DateTime ManufacturingDate { get; set; }
 
         string ToString();
